pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "open-argondb"
version = "0.1.0"
description = "Enterprise-equivalent features for ArangoDB Community Edition — CDC, audit, encryption, agent memory, and more."
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
authors = [{ name = "aarjay" }]
keywords = ["arangodb", "agent-memory", "cdc", "audit", "vector-search", "graph"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Topic :: Database",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "python-arango>=8.0",
    "numpy>=1.24",
]

[project.optional-dependencies]
embeddings = ["sentence-transformers>=2.2"]
events-redis = ["redis>=5.0"]
events-nats = ["nats-py>=2.0"]
mcp = ["mcp>=1.0"]
all = [
    "sentence-transformers>=2.2",
    "redis>=5.0",
    "nats-py>=2.0",
    "mcp>=1.0",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.0",
    "ruff>=0.3",
    "mypy>=1.8",
]

[project.urls]
Homepage = "https://github.com/aarjay/open-argondb"
Repository = "https://github.com/aarjay/open-argondb"
Issues = "https://github.com/aarjay/open-argondb/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/open_argondb"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.10"
strict = true
